requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
google-genai==1.0.0
flask
flask-sqlalchemy
sift-stack-py
sqlalchemy
